+++ source/extensions/omni.isaac.quadruped/omni/isaac/quadruped/tests/test_a1.py
@@ -206,7 +206,7 @@ class TestA1(omni.kit.test.AsyncTestCase):
         return
 
     def on_physics_step(self, step_size):
-        if self._a1 and self._a1._handle:
+        if self._a1:
             self._a1.advance(
                 dt=step_size, goal=self._base_command, path_follow=self._path_follow, auto_start=self._auto_start
             )