+++ source/extensions/omni.isaac.tests/omni/isaac/tests/tests/test_jetbot.py
@@ -20,10 +20,9 @@ import omni.graph.core as og
 #   For most things refer to unittest docs: https://docs.python.org/3/library/unittest.html
 import omni.kit.test
 from omni.isaac.core import World
+from omni.isaac.core.articulations.articulation import Articulation
 from omni.isaac.core.utils.prims import delete_prim
 from omni.isaac.core.utils.stage import open_stage_async
-from omni.isaac.dynamic_control import _dynamic_control
-from omni.isaac.dynamic_control import utils as dc_utils
 from omni.isaac.nucleus import get_assets_root_path_async
 
 from .robot_helpers import init_robot_sim, setup_robot_og
@@ -33,8 +32,6 @@ from .robot_helpers import init_robot_sim, setup_robot_og
 class TestJetBot(omni.kit.test.AsyncTestCase):
     # Before running each test
     async def setUp(self):
-        self.dc = _dynamic_control.acquire_dynamic_control_interface()
-
         self._assets_root_path = await get_assets_root_path_async()
         if self._assets_root_path is None:
             carb.log_error("Could not find Isaac Sim assets folder")
@@ -79,20 +76,20 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
 
         # get the dofbot
-        self.ar = self.dc.get_articulation("/jetbot")
-        self.chassis = self.dc.get_articulation_root_body(self.ar)
-        self.starting_pos = np.array(self.dc.get_rigid_body_pose(self.chassis).p)
-
-        self.wheel_left = self.dc.find_articulation_dof(self.ar, "left_wheel_joint")
-        self.wheel_right = self.dc.find_articulation_dof(self.ar, "right_wheel_joint")
+        self.ar = Articulation("/jetbot")
+        self.ar._articulation_view.initialize()
+        self.starting_pos, _ = self.ar.get_world_pose()
+        left_wheel_joint_idx = self.ar._articulation_view.get_dof_index("left_wheel_joint")
+        right_wheel_joint_idx = self.ar._articulation_view.get_dof_index("right_wheel_joint")
+        self.ar._articulation_view.set_joint_velocity_targets(
+            velocities=np.array([1.0, 1.0]), joint_indices=[left_wheel_joint_idx, right_wheel_joint_idx]
+        )
 
         # move the jetbot
-        self.dc.set_dof_velocity_target(self.wheel_left, 1.0)
-        self.dc.set_dof_velocity_target(self.wheel_right, 1.0)
+        for i in range(60):
+            await omni.kit.app.get_app().next_update_async()
 
-        await dc_utils.simulate(1, self.dc, self.ar)
-
-        self.current_pos = np.array(self.dc.get_rigid_body_pose(self.chassis).p)
+        self.current_pos, _ = self.ar.get_world_pose()
 
         delta = np.linalg.norm(self.current_pos - self.starting_pos)
         print("Diff is ", delta)
@@ -111,7 +108,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/jetbot")
+        await init_robot_sim("/jetbot")
 
         for x in range(1, 5):
             forward_velocity = x * 0.10
@@ -142,7 +139,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/jetbot")
+        await init_robot_sim("/jetbot")
         for x in range(1, 5):
             self.my_world.play()
             await omni.kit.app.get_app().next_update_async()
@@ -178,7 +175,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
             self.my_world.play()
             await omni.kit.app.get_app().next_update_async()
 
-            await init_robot_sim(self.dc, "/jetbot")
+            await init_robot_sim("/jetbot")
 
             angular_velocity = 0.6 * x
             og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:angularVelocity").set(
@@ -206,7 +203,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/jetbot")
+        await init_robot_sim("/jetbot")
         forward_velocity = -0.1
         angular_velocity = -0.5
         og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(forward_velocity)
@@ -233,7 +230,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
     #     self.my_world.play()
     #     await omni.kit.app.get_app().next_update_async()
 
-    #     await init_robot_sim(self.dc, "/jetbot")
+    #     await init_robot_sim("/jetbot")
     #     l_wheel = self.dc.get_rigid_body("/jetbot/left_wheel")
 
     #     forward_velocity = 0.2
@@ -253,7 +250,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
     #             self.my_world.play()
     #             await omni.kit.app.get_app().next_update_async()
 
-    #             await init_robot_sim(self.dc, "/jetbot")
+    #             await init_robot_sim("/jetbot")
     #             # l_wheel = self.dc.get_rigid_body("/jetbot/left_wheel")
 
     #             # wait until const velocity reached