+++ source/extensions/omni.isaac.utils/python/scripts/commands.py
@@ -12,7 +12,6 @@ import asyncio
 import carb
 import omni.kit.commands
 import omni.kit.utils
-from omni.isaac.dynamic_control import _dynamic_control
 from omni.isaac.utils._isaac_utils import transforms
 from pxr import Sdf
 
@@ -41,7 +40,6 @@ class IsaacSimSpawnPrim(omni.kit.commands.Command):
             if name != "self":
                 setattr(self, f"_{name}", value)
         pass
-        self._dc = _dynamic_control.acquire_dynamic_control_interface()
         self._stage = omni.usd.get_context().get_stage()
         self._context = omni.usd.get_context()
         pass
@@ -51,7 +49,6 @@ class IsaacSimSpawnPrim(omni.kit.commands.Command):
         self._prim.GetReferences().AddReference(self._usd_path)
         if self._translation is not None and self._rotation is not None:
             transforms.set_transform(
-                self._dc,
                 self._context.get_stage_id(),
                 str(self._prim.GetPath()),
                 tuple(self._translation),
@@ -84,14 +81,13 @@ class IsaacSimTeleportPrim(omni.kit.commands.Command):
         for name, value in vars().items():
             if name != "self":
                 setattr(self, f"_{name}", value)
-        self._dc = _dynamic_control.acquire_dynamic_control_interface()
         self._context = omni.usd.get_context()
         pass
 
     def do(self) -> bool:
         if self._translation is not None and self._rotation is not None:
             transforms.set_transform(
-                self._dc, self._context.get_stage_id(), str(self._prim_path), self._translation, self._rotation
+                self._context.get_stage_id(), str(self._prim_path), self._translation, self._rotation
             )
         return True
         pass
@@ -118,13 +114,12 @@ class IsaacSimScalePrim(omni.kit.commands.Command):
         for name, value in vars().items():
             if name != "self":
                 setattr(self, f"_{name}", value)
-        self._dc = _dynamic_control.acquire_dynamic_control_interface()
         self._context = omni.usd.get_context()
         pass
 
     def do(self) -> bool:
         if self._scale is not None:
-            transforms.set_scale(self._dc, self._context.get_stage_id(), str(self._prim_path), self._scale)
+            transforms.set_scale(self._context.get_stage_id(), str(self._prim_path), self._scale)
         return True
         pass