+++ source/extensions/omni.isaac.tests/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.16.2] - 2023-03-07
+### Changed
+- Removed the usage of the deprecated dynamic_control extension 
+
 ## [1.16.1] - 2023-02-02
 ### Changed
 - Updated path to the nucleus extension